requirements.txt
openai
eel
mysql-connector-python
opencv-python
qrcode[pil]